Clean up and change int to int16

+++ Math/Vector.h
@@ -4,8 +4,7 @@
 #include <malloc.h>
 #include <string.h>
 #include <stdarg.h>
-
-//#include <cuda.h>
+#include <stdint.h>
 
 
 //Voodoo Magick
@@ -28,7 +27,7 @@
 //And a dimension, or number of components.
 typedef struct Vector
 {
-	int dimension;
+	uint16_t  dimension;
 	float* components;
 }Vector;
 
@@ -52,7 +51,7 @@ extern const Vector Vector_E3;
 // Allocates a Vector
 //
 //Returns:
-//	A pointer to the newly allocated Vector. 
+//	A pouint16_t er to the newly allocated Vector. 
 //	Vector is not initialized yet and has no components.
 Vector* Vector_Allocate();
 
@@ -61,7 +60,7 @@ Vector* Vector_Allocate();
 //
 //Parameters:
 //	vec: The Vector to initialize
-void Vector_Initialize(Vector* vec, int dim);
+void Vector_Initialize(Vector* vec, uint16_t dim);
 
 ///
 //Frees the memory taken by a Vector
@@ -77,7 +76,7 @@ void Vector_Free(Vector* vec);
 //	dest: The destination of the copy
 //	src: the Vector to copy
 //	dim: The dimension of the Vector being copied
-void Vector_CopyArray(float* dest, const float* src, const int dim);
+void Vector_CopyArray(float* dest, const float* src, const uint16_t  dim);
 //Checks for errors then calls Vector_CopyArray
 void Vector_Copy(Vector* dest, const Vector* src);
 
@@ -87,7 +86,7 @@ void Vector_Copy(Vector* dest, const Vector* src);
 //Parameters:
 //	vec: The Vector being zeroed
 //	dim: The number of components the Vector has
-void Vector_ZeroArray(float* vec, const int dim);
+void Vector_ZeroArray(float* vec, const uint16_t  dim);
 //Calls Vector_ZeroArray
 void Vecor_Zero(Vector* vec);
 
@@ -97,7 +96,7 @@ void Vecor_Zero(Vector* vec);
 //Parameters:
 //	vec: The vector to find the mag squared of
 //	dim: The number of components in the vector
-float Vector_GetMagSqFromArray(const float* vec, const int dim);
+float Vector_GetMagSqFromArray(const float* vec, const uint16_t  dim);
 float Vector_GetMagSq(const Vector* vec);
 
 ///
@@ -106,7 +105,7 @@ float Vector_GetMagSq(const Vector* vec);
 //Parameters:
 //	vec: The Vector to find the magnitude of
 //	dim: the number of components in the Vector
-float Vector_GetMagFromArray(const float* vec, const int dim);
+float Vector_GetMagFromArray(const float* vec, const uint16_t  dim);
 float Vector_GetMag(const Vector* vec);
 
 ///
@@ -115,7 +114,7 @@ float Vector_GetMag(const Vector* vec);
 //Parameters:
 //	vec: the components to normalize
 //	dim: the dimension of the Vector
-void Vector_NormalizeArray(float* vec, const int dim);
+void Vector_NormalizeArray(float* vec, const uint16_t  dim);
 void Vector_Normalize(Vector* vec);
 
 ///
@@ -128,7 +127,7 @@ void Vector_Normalize(Vector* vec);
 //
 //Returns:
 //	Float representing value of the dot product
-float Vector_DotProductArray(const float* vec1, const float* vec2, const int dim);
+float Vector_DotProductArray(const float* vec1, const float* vec2, const uint16_t  dim);
 //Checks for errors then calls Vector_DotProductArray
 //Returns 0 on error
 float Vector_DotProduct(const Vector* vec1, const Vector* vec2);
@@ -140,7 +139,7 @@ float Vector_DotProduct(const Vector* vec1, const Vector* vec2);
 //	dest: The destination of the cross product result
 //	dim: The dimension of the resulting vector
 //	vectors: an array of dim - 1 vectors of dimension dim
-void Vector_CrossProductArray(float* dest, const unsigned int dim, float** vectors);
+void Vector_CrossProductArray(float* dest, const uint16_t  dim, float** vectors);
 ///
 //Checks for errors then calls CVector_CrossProductArray
 //
@@ -159,7 +158,7 @@ void Vector_CrossProduct(Vector* dest, ...);
 //
 //returns:
 //	Angle in radians
-float Vector_GetAngleArray(const float* vec1, const float* vec2, const int dim);
+float Vector_GetAngleArray(const float* vec1, const float* vec2, const uint16_t  dim);
 //Checks for errors then calls CVector_GetAngleArray
 float Vector_GetAngle(const Vector* vec1, const Vector* vec2);
 
@@ -171,9 +170,9 @@ float Vector_GetAngle(const Vector* vec1, const Vector* vec2);
 //	dest: the Vector getting incremented
 //	src: The addend Vector, or the Vector being added to the destination
 //	dim: The number of components in the Vectors
-void Vector_IncrementArray(float* dest, const float* src, const int dim);
+void Vector_IncrementArray(float* dest, const float* src, const uint16_t  dim);
 void Vector_Increment(Vector* dest, const Vector* src);
-void Vector_DecrementArray(float* dest, const float* src, const int dim);
+void Vector_DecrementArray(float* dest, const float* src, const uint16_t  dim);
 void Vector_Decrement(Vector* dest, const Vector* src);
 
 ///
@@ -183,7 +182,7 @@ void Vector_Decrement(Vector* dest, const Vector* src);
 //	vec: The Vector to be scaled
 //	scaleValue: the scale factor
 //	dim: the dimension of the Vector being scaled
-void Vector_ScaleArray(float* vec, const float scaleValue, const int dim);
+void Vector_ScaleArray(float* vec, const float scaleValue, const uint16_t  dim);
 void Vector_Scale(Vector* vec, const float scaleValue);
 
 ///
@@ -194,7 +193,7 @@ void Vector_Scale(Vector* vec, const float scaleValue);
 //	src: the initial Vector to be scaled
 //	scaleValue: the amount by which to scale the initial Vector
 //	dim: The number of components in the initial Vector
-void Vector_GetScalarProductFromArray(float* dest, const float* src, const float scaleValue, const int dim);
+void Vector_GetScalarProductFromArray(float* dest, const float* src, const float scaleValue, const uint16_t  dim);
 void Vector_GetScalarProduct(Vector* dest, const Vector* src, const float scaleValue);
 
 ///
@@ -205,9 +204,9 @@ void Vector_GetScalarProduct(Vector* dest, const Vector* src, const float scaleV
 //	vec1: the first addend Vector
 //	vec2: the second addend Vector
 //	dim: the dimension of the Vectors being added
-void Vector_AddArray(float* dest, const float* vec1, const float* vec2, const int dim);
+void Vector_AddArray(float* dest, const float* vec1, const float* vec2, const uint16_t  dim);
 void Vector_Add(Vector* dest, const Vector* vec1, const Vector* vec2);
-void Vector_SubtractArray(float* dest, const float* vec1, const float* vec2, const int dim);
+void Vector_SubtractArray(float* dest, const float* vec1, const float* vec2, const uint16_t  dim);
 void Vector_Subtract(Vector* dest, const Vector* vec1, const Vector* vec2);
 
 ///
@@ -218,7 +217,7 @@ void Vector_Subtract(Vector* dest, const Vector* vec1, const Vector* vec2);
 //	vec1: The vector being projected
 //	vec2: The vector representing the projection axis
 //	dim: the dimension of the vectors
-void Vector_GetProjectionArray(float* dest, const float* vec1, const float* vec2, const int dim);
+void Vector_GetProjectionArray(float* dest, const float* vec1, const float* vec2, const uint16_t  dim);
 void Vector_GetProjection(Vector* dest, const Vector* vec1, const Vector* vec2);
 
 ///
@@ -228,7 +227,7 @@ void Vector_GetProjection(Vector* dest, const Vector* vec1, const Vector* vec2);
 //	vec1: The vector being projected
 //	vec2: The vector representing the projection axis
 //	dim: The dimension of the vectors
-void Vector_ProjectArray(float* vec1, const float* vec2, const int dim);
+void Vector_ProjectArray(float* vec1, const float* vec2, const uint16_t  dim);
 void Vector_Project(Vector* vec1, const Vector* vec2);
 
 ///
@@ -237,10 +236,9 @@ void Vector_Project(Vector* vec1, const Vector* vec2);
 //Parameters:
 //	vec: the Vector to print
 //	dim: the number of components in the Vector
-void Vector_PrintTransposeArray(const float* vec, const int dim);
+void Vector_PrintTransposeArray(const float* vec, const uint16_t  dim);
 void Vector_PrintTranspose(const Vector* vec);
-void Vector_PrintArray(const float* vec, const int dim);
+void Vector_PrintArray(const float* vec, const uint16_t  dim);
 void Vector_Print(const Vector* vec);
 
 #endif
-
